wip(today): quick-reschedule writes dueDate only
Refs: lean.gem
